p4raw-id: //depot/perlio@16064

+++ ext/IO/poll.c
@@ -58,7 +58,7 @@ again:
     FD_ZERO(&wfd);
     FD_ZERO(&efd);
 
-    for(i = 0 ; i < nfds ; i++) {
+    for(i = 0 ; i < (int)nfds ; i++) {
 	int events = fds[i].events;
 	int fd = fds[i].fd;
 
@@ -105,7 +105,7 @@ again:
 
     count = 0;
 
-    for(i = 0 ; i < nfds ; i++) {
+    for(i = 0 ; i < (int)nfds ; i++) {
 	int revents = (fds[i].events & POLL_EVENTS_MASK);
 	int fd = fds[i].fd;
